refactor: Kakao로그인 관련 코드 정리
